added test time manipulation
